Добавление interceptor Изменение температуры из лонга в флоат

+++ app/src/main/java/com/example/weatherapp/data/network/dto/DayWeatherDto.kt
@@ -3,6 +3,6 @@ package com.example.weatherapp.data.network.dto
 import com.google.gson.annotations.SerializedName
 
 data class DayWeatherDto(
-    @SerializedName("avgtemp_c") val tempC: Long,
+    @SerializedName("avgtemp_c") val tempC: Float,
     @SerializedName("condition") val conditionDto: ConditionDto,
 )
