+++ app/src/main/java/com/example/weatherapp/data/network/dto/WeatherDto.kt
@@ -4,6 +4,6 @@ import com.google.gson.annotations.SerializedName
 
 data class WeatherDto(
     @SerializedName("last_updated_epoch") val date: Long,
-    @SerializedName("temp_c") val tempC: Long,
+    @SerializedName("temp_c") val tempC: Float,
     @SerializedName("condition") val conditionDto: ConditionDto,
 )
